Update README.md
Abstract— Nadi Pariksha is a non-invasive pulse diagnosis 
method for assessing health of person in Ayurveda. The disease 
caused can be related to imbalance in any of three doshas i.e. 
vatha, pittha and kapha. The health status of the person is 
examined by Ayurvedic physician by feeling palpation from 
three fingers (index, middle and ring) placed on the radial 
artery for vatha, pittha and kapha respectively. In present 
research work the design and development of the pulse 
diagnosis system was carried out, in which 
Photoplethysmography (PPG) sensor was interfaced with 
Arduino was used to acquire wrist pulse signals from subjects 
in the form of data sets. Matlab software was used in signal 
processing after performing frequency domain analysis using 
power spectrum to design a filter and feature extraction was 
performed on these signals. Algorithms for three parameters 
were designed related to pulse sensors namely pulse wave 
velocity, augmentation index and reflectivity index. Data sets 
were created with the help of hundred volunteers. After 
performing statistical analysis on these data sets a standard 
signal was established. Finally this experimental setup was 
used to diagnose patients with symptoms associated with 
imbalance in Tridosha to look for deviations from the standard 
vatha pittha kapha signal and these results were compared 
with theoretical results as per Ayurveda. Based on this 
research we came up with a conclusion with respect to the 
system designed.